requirements.txt
fastapi
uvicorn
pydantic
pyjwt
passlib
sqlalchemy
python-jose
alembic
authlib
psycopg2
pydantic[email]
python-dotenv
twilio
python-multipart
boto3
pydantic_settings